+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## NEXT
+## 0.0.4
 
 - Adds `interceptCallback` parameter to allow for custom interception of the callback URL
+- Adds `isAuthenticated` method to check if the user is authenticated
 
 ## 0.0.3